// src/resources/coffee-shop/statistics/types/statistics-percentages.ts
export interface StatisticsPercentages {
    salaryPercent: number;
    costPercent: number;
    writeOffPercent: number;
    inventoryShortagePercent: number;
    inventorySurplusPercent: number;
    inventoryNetAdjustmentPercent: number;
    cashPercent: number;
    terminalPercent: number;
    acquiringPercent: number;
}